pyproject.toml
[tool.poetry]
name = "python-web-hw14"
version = "0.1.0"
description = ""
authors = ["Val <yermakvaleriy@gmail.com>"]
readme = "README.txt"

[tool.poetry.dependencies]
python = "^3.11"
unicorn = "^2.0.1.post1"
sqlalchemy = "^2.0.29"
fastapi = "^0.110.0"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.6.4"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
asyncpg = "^0.29.0"
bcrypt = "3.2.2"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"
redis = "^5.0.3"
limiter = "^0.5.0"
fastapi-limiter = "^0.1.6"
jinja2 = "^3.1.3"
cloudinary = "^1.39.1"
sphinx = "^7.2.6"
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"



[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["tests",]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"